requirements.txt
# Restaurant Reservation System - Python Dependencies
# MIT400 Assessment 2

# Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1

# Database
mysql-connector-python==8.1.0
PyMySQL==1.1.0
psycopg2-binary==2.9.9

# Security
bcrypt==4.0.1
Werkzeug==2.3.7

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0

# Additional utilities
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==22.0.0
